process rules
git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/branches/johns_perma_branch@1527 72d1a826-0e72-11de-b5bf-434079a3519f

+++ mst-platform/src/java/xc/mst/action/processingDirective/DeleteProcessingDirective.java
@@ -19,8 +19,11 @@ import org.apache.log4j.Logger;
 import xc.mst.action.BaseActionSupport;
 import xc.mst.bo.processing.Job;
 import xc.mst.bo.processing.ProcessingDirective;
+import xc.mst.bo.provider.Provider;
+import xc.mst.bo.service.Service;
 import xc.mst.constants.Constants;
 import xc.mst.dao.DatabaseConfigException;
+import xc.mst.repo.Repository;
 import xc.mst.utils.MSTConfiguration;
 
 /**
@@ -92,21 +95,41 @@ public class DeleteProcessingDirective extends BaseActionSupport
             //        if all successors of the of the records from the input side are marked deleted
             //        then   ---->  the processing rule will be deleted immediately.
             if (!isProcessingDirectiveInUse(tempProcDir)) {
-                getProcessingDirectiveService().deleteProcessingDirective(tempProcDir);
-                deleted = true;
-                return SUCCESS;
+            	
+            	if (!hasNonDeletedSourceRecords(tempProcDir)) {
+            		if (!hasNonDeletedDestinationRecords(tempProcDir)) {
+                        getProcessingDirectiveService().deleteProcessingDirective(tempProcDir);
+                        deleted = true;
+                        return SUCCESS;
+            		}
+                	else {
+                        deleted = false; // this flag will be used to decide whether to show the 2nd dialog.
+
+                        Object[] messageArguments = {getOutputRepository(tempProcDir).getName()}; 
+                        message = MSTConfiguration.getMSTString("message.processingRuleNonDeletedDestRecords", messageArguments);
+LOG.debug("**** action error: message set to: "+message);                        
+                        return INPUT;   //TODO
+                	}
+            	}
+            	else {
+                    deleted = false; // this flag will be used to decide whether to show the 2nd dialog.
+
+                    Object[] messageArguments = {getInputRepository(tempProcDir).getName()}; 
+                    message = MSTConfiguration.getMSTString("message.processingRuleNonDeletedSrcRecords", messageArguments);
+LOG.debug("**** action error: message set to: "+message);                        
+                    
+                    return INPUT;   //TODO
+            	}	
             }
-//            else if () {
-//            		// need to create code to determine if records marked deleted or if no records exist.
-//            }
-            else {
+        	else {
                 deleted = false; // this flag will be used to decide whether to show the 2nd dialog.
 
                 Object[] messageArguments = {tempProcDir.getId()}; 
                 message = MSTConfiguration.getMSTString("message.processingRuleInUse", messageArguments);
-                
+ LOG.debug("**** action error: message set to: "+message);                        
+               
                 return INPUT;   //TODO
-            }
+        	}
         }
         catch(DatabaseConfigException e)
         {
@@ -115,6 +138,48 @@ public class DeleteProcessingDirective extends BaseActionSupport
             errorType = "error";
             return SUCCESS;
         }
+    }
+
+    private Repository getOutputRepository(ProcessingDirective tempProcDir) {
+    	Service outputS = tempProcDir.getService();  // the service used to produce the output
+    	Repository outgoingRepo = getRepository(outputS);
+    	
+    	return  outgoingRepo;
+    }
+
+    private Repository getInputRepository(ProcessingDirective tempProcDir) {
+    	Provider provider = tempProcDir.getSourceProvider();
+    	Repository incomingRepo = getRepositoryService().getRepository(provider);
+    	
+    	return incomingRepo;
+    }
+    
+    private Repository getRepository(Service outputS) {
+    	List<Repository> list =  getRepositoryService().getAll();
+    	for (Repository r: list) {
+    		if (r.getService().equals(outputS)) {
+    			return r;
+    		}
+    	}
+    	return null;
+    }
+    
+    private boolean hasNonDeletedSourceRecords(ProcessingDirective tempProcDir){
+    	return hasNonDeletedRecords(getInputRepository(tempProcDir));
+    }
+
+    private boolean hasNonDeletedDestinationRecords(ProcessingDirective tempProcDir){
+    	return hasNonDeletedRecords(getOutputRepository(tempProcDir));
+    }
+    
+    private boolean hasNonDeletedRecords(Repository r){
+    	if (!getRepositoryDAO().exists(r.getName())) {
+    		return false;
+    	}
+    	else if (getRepositoryDAO().hasOnlyRecordsOfStatus(r.getName(), 'D')) {
+    		return false;
+    	}
+    	return true;
     }
 
     private boolean isProcessingDirectiveInUse(ProcessingDirective tempProcDir)
